Support flow control Task.

+++ aliyun-java-sdk-airec/src/main/java/com/aliyuncs/airec/model/v20201126/DescribeSyncReportOutliersRequest.java
@@ -29,9 +29,9 @@ public class DescribeSyncReportOutliersRequest extends RoaAcsRequest<DescribeSyn
 
 	private String levelType;
 
-	private Integer endTime;
+	private Long endTime;
 
-	private Integer startTime;
+	private Long startTime;
 
 	private String type;
 
@@ -68,22 +68,22 @@ public class DescribeSyncReportOutliersRequest extends RoaAcsRequest<DescribeSyn
 		}
 	}
 
-	public Integer getEndTime() {
+	public Long getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Integer endTime) {
+	public void setEndTime(Long endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
 		}
 	}
 
-	public Integer getStartTime() {
+	public Long getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Integer startTime) {
+	public void setStartTime(Long startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
