+++ aliyun-java-sdk-airec/src/main/java/com/aliyuncs/airec/model/v20201126/ListDashboardMetricsRequest.java
@@ -31,9 +31,9 @@ public class ListDashboardMetricsRequest extends RoaAcsRequest<ListDashboardMetr
 
 	private String metricQuery;
 
-	private Integer endTime;
+	private Long endTime;
 
-	private Integer startTime;
+	private Long startTime;
 
 	private String metricView;
 	public ListDashboardMetricsRequest() {
@@ -79,22 +79,22 @@ public class ListDashboardMetricsRequest extends RoaAcsRequest<ListDashboardMetr
 		}
 	}
 
-	public Integer getEndTime() {
+	public Long getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Integer endTime) {
+	public void setEndTime(Long endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
 		}
 	}
 
-	public Integer getStartTime() {
+	public Long getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Integer startTime) {
+	public void setStartTime(Long startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
