+++ aliyun-java-sdk-airec/src/main/java/com/aliyuncs/airec/model/v20201126/QueryDataMessageStatisticsRequest.java
@@ -29,11 +29,11 @@ public class QueryDataMessageStatisticsRequest extends RoaAcsRequest<QueryDataMe
 
 	private String messageSource;
 
-	private Integer endTime;
+	private Long endTime;
 
 	private String userType;
 
-	private Integer startTime;
+	private Long startTime;
 
 	private String userId;
 
@@ -84,11 +84,11 @@ public class QueryDataMessageStatisticsRequest extends RoaAcsRequest<QueryDataMe
 		}
 	}
 
-	public Integer getEndTime() {
+	public Long getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Integer endTime) {
+	public void setEndTime(Long endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
@@ -106,11 +106,11 @@ public class QueryDataMessageStatisticsRequest extends RoaAcsRequest<QueryDataMe
 		}
 	}
 
-	public Integer getStartTime() {
+	public Long getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Integer startTime) {
+	public void setStartTime(Long startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
